refactor(skill-progression): extract Layer 4 EMA engine into _ema_engine.py
Move calculate_tournament_skill_contribution and
compute_single_tournament_skill_delta verbatim from
skill_progression_service.py (shim) into the sub-package at
app/services/skill_progression/_ema_engine.py.

Both symbols are re-exported from the package __init__.py and imported
back into the shim.  Layer 5 functions (get_skill_timeline, get_skill_audit,
get_avg_skill_level_checkpoints) remain in the shim so their LOAD_GLOBAL
patch interception is unaffected.

Test patch-path update (correctness, not cleanup):
  _ema_engine.py has its own __dict__, so tests that call ctsc/cstsd
  directly and patch their internal dependencies must target
  "app.services.skill_progression._ema_engine.*" instead of the shim.
  Added _BASE_EMA + 6 _PATCH_*_EMA constants; updated 15 test methods
  in TestCalculateTournamentSkillContribution,
  TestComputeSingleTournamentSkillDelta, and the C/D sub-section of
  TestEdgeCasesBeyondBasic.  Layer 5 test methods unchanged.

Zero-behaviour-change: 7186/7186 unit tests pass (21 xfailed, 1 xpassed —
identical profile to main).
